requirements.txt
fastapi
uvicorn
opencv-python
pytesseract
Pillow
python-multipart
pytest
httpx
